extract some methods from the pdf generator to a common file
